+++ src/libasr/CMakeLists.txt
@@ -33,6 +33,7 @@ set(SRC
     pass/param_to_const.cpp
     pass/do_loops.cpp
     pass/for_all.cpp
+    pass/while_else.cpp
     pass/global_stmts.cpp
     pass/select_case.cpp
     pass/init_expr.cpp
