+++ src/libasr/pass/pass_list_expr.cpp
@@ -92,7 +92,7 @@ public:
         loop_body.push_back(al, loop_stmt);
 
         loop_stmt = ASRUtils::STMT(ASR::make_WhileLoop_t(
-            al, loc, nullptr, loop_test, loop_body.p, loop_body.n));
+            al, loc, nullptr, loop_test, loop_body.p, loop_body.n, nullptr, 0));
         body.push_back(al, loop_stmt);
     }
